requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.3
